enforce last argument (offset) of mmap function to be Int64, hoping to pass Win32 test

+++ src/SnpArrays.jl
@@ -49,13 +49,13 @@ function SnpArray(
     if bits(bedheader[3]) == "00000001"
       # SNP-major
       plinkbits = Mmap.mmap(plinkBedfile, BitArray{3},
-        (2, 4ceil(Int, 0.25people), snps), 3)
+        (2, 4ceil(Int, 0.25people), snps), Int64(3))
       A1 = copy!(A1, slice(plinkbits, 1, 1:people, :))
       A2 = copy!(A2, slice(plinkbits, 2, 1:people, :))
     else
       # individual-major
       snpbits = Mmap.mmap(plinkBedfile, BitArray{3},
-        (2, 4ceil(Int, 0.25snps), people), 3)
+        (2, 4ceil(Int, 0.25snps), people), Int64(3))
       A1 = copy!(A1, slice(plinkbits, 1, 1:people, :)')
       A2 = copy!(A2, slice(plinkbits, 2, 1:people, :)')
     end
